Enable and fix multiboot-preset framebuffer sizing

+++ base/usr/include/kernel/multiboot.h
@@ -27,24 +27,32 @@ struct multiboot
 	uint32_t cmdline;
 	uint32_t mods_count;
 	uint32_t mods_addr;
+
 	uint32_t num;
 	uint32_t size;
 	uint32_t addr;
 	uint32_t shndx;
+
 	uint32_t mmap_length;
 	uint32_t mmap_addr;
+
 	uint32_t drives_length;
 	uint32_t drives_addr;
+
 	uint32_t config_table;
+
 	uint32_t boot_loader_name;
+
 	uint32_t apm_table;
+
 	uint32_t vbe_control_info;
 	uint32_t vbe_mode_info;
-	uint32_t vbe_mode;
-	uint32_t vbe_interface_seg;
-	uint32_t vbe_interface_off;
-	uint32_t vbe_interface_len;
-	uint32_t framebuffer_addr;
+	uint16_t vbe_mode;
+	uint16_t vbe_interface_seg;
+	uint16_t vbe_interface_off;
+	uint16_t vbe_interface_len;
+
+	uint64_t framebuffer_addr;
 	uint32_t framebuffer_pitch;
 	uint32_t framebuffer_width;
 	uint32_t framebuffer_height;
